make the Update command work the way we want it to

+++ src/ripple/Nuget/NugetService.cs
@@ -1,11 +1,15 @@
 using System;
 using System.Collections.Generic;
+using System.Xml;
+using FubuCore.CommandLine;
+using FubuCore.Util;
 using NuGet;
 using NuGet.Common;
 using ripple.Local;
 using ripple.Model;
 using Console = NuGet.Common.Console;
 using System.Linq;
+using FubuCore;
 
 namespace ripple.Nuget
 {
@@ -18,6 +22,7 @@ namespace ripple.Nuget
         private readonly PackageManager _packageManager;
         private readonly Console _console;
         private readonly DefaultPackagePathResolver _pathResolver;
+        private Cache<NugetDependency, IPackage> _packages;
 
         public NugetService(Solution solution)
         {
@@ -37,6 +42,12 @@ namespace ripple.Nuget
             _packageManager = new PackageManager(_sourceRepository, _pathResolver, _fileSystem, _localRepository){
                 Logger = _console
             };
+
+            _packages = new Cache<NugetDependency, IPackage>(dep =>
+            {
+                Install(dep);
+                return _sourceRepository.FindPackage(dep.Name, dep.Version);
+            });
         }
 
         // TODO -- make this allow changes later.  Custom repo's etc.
@@ -46,46 +57,79 @@ namespace ripple.Nuget
 
         public NugetDependency GetLatest(string nugetName)
         {
-            var package = _remoteRepository.Search("FubuCore").Where(x => x.IsLatestVersion).FirstOrDefault();
+            var package = _remoteRepository.Search(nugetName).Where(x => x.Id == nugetName && x.IsLatestVersion).FirstOrDefault();
             return package == null ? null : new NugetDependency(package.Id, package.Version.ToString());
         }
 
         public void Install(NugetDependency dependency)
         {
             var version = new Version(dependency.Version);
-            _packageManager.InstallPackage(dependency.Name, version);
-        }
-
-        public void Restore(NugetDependency dependency)
-        {
-            var version = new Version(dependency.Version);
             _packageManager.InstallPackage(dependency.Name, version, true);
         }
 
+
         public void RemoveFromFileSystem(NugetDependency dependency)
         {
+            ConsoleWriter.PrintHorizontalLine();
+            ConsoleWriter.Write(ConsoleColor.Cyan, "Removing " + dependency);
+
+
             var package = _localRepository.FindPackage(dependency.Name, dependency.Version);
-            _localRepository.RemovePackage(package);
+
+            if (package != null) _localRepository.RemovePackage(package);
         }
 
         public void Update(Project project, IEnumerable<NugetDependency> dependencies)
         {
-            var projectSystem = new MSBuildProjectSystem(project.ProjectFile);
-            var sharedPackageRepository = new SharedPackageRepository(_pathResolver, _fileSystem);
-            var projectRepository = new PackageReferenceRepository(_fileSystem, sharedPackageRepository);
+            var projectManager = buildProjectManager(project);
 
-            var projectManager = new ProjectManager(_sourceRepository, _pathResolver, projectSystem, projectRepository){
-                Logger = _console
-            };
+            ConsoleWriter.PrintHorizontalLine();
+            ConsoleWriter.Write(ConsoleColor.Cyan, "Updating project " + project.ProjectName);
+
+
 
             dependencies.Each(dep =>
             {
-                Install(dep);
+                ConsoleWriter.PrintHorizontalLine();
+                ConsoleWriter.Write(ConsoleColor.Cyan, "  -- to " + dep);
 
-                var version = new Version(dep.Version);
-                var package = _localRepository.FindPackage(dep.Name, version);
-                
-                projectManager.AddPackageReference(package, false);
+                projectManager.AddPackageReference(_packages[dep], true);
+            });
+        }
+
+        private ProjectManager buildProjectManager(Project project)
+        {
+            var projectSystem = new MSBuildProjectSystem(project.ProjectFile);
+            var fileSystem = new PhysicalFileSystem(project.ProjectFile.ParentDirectory());
+            var sharedPackageRepository = new SharedPackageRepository(_pathResolver, fileSystem);
+            var projectRepository = new PackageReferenceRepository(fileSystem, sharedPackageRepository);
+
+            return new ProjectManager(_sourceRepository, _pathResolver, projectSystem, projectRepository){
+                Logger = _console
+            };
+        }
+
+        public void RemoveFromProject(Project project, IEnumerable<NugetDependency> dependencies)
+        {
+            var projectManager = buildProjectManager(project);
+            var document = new XmlDocument();
+            document.Load(project.PackagesFile());
+            
+
+            dependencies.Each(dep =>
+            {
+                ConsoleWriter.Write("Trying to remove {0} from {1}", dep, project.ProjectName);
+                projectManager.RemovePackageReference(_packages[dep], true, false);
+
+                var element =
+                    document.DocumentElement.SelectSingleNode("package[@id='{0}' and @version='{1}']".ToFormat(
+                        dep.Name, dep.Version));
+
+                if (element != null)
+                {
+                    document.DocumentElement.RemoveChild(element);
+                    document.Save(project.PackagesFile());
+                }
             });
         }
     }
